JIP-26 REFACTOR: interface 분리
