+++ backend/src/books/books.type.ts
@@ -1,0 +1,13 @@
+export interface SearchBookInfoQuery {
+    query: string;
+    sort: string;
+    page: string;
+    limit: string;
+    category: string;
+};
+
+
+export interface SortInfoType { 
+    sort: string;
+    limit: string;
+};
